+++ inductiva/molecules/__init__.py
@@ -1,4 +1,3 @@
 #pylint: disable=missing-module-docstring
-from .simulators import GROMACS
 from .scenarios import ProteinSolvation
 from .scenarios import MDWaterBox
