+++ inductiva/molecules/scenarios/protein_solvation/protein_solvation.py
@@ -6,7 +6,7 @@ import os
 import shutil
 
 from inductiva import tasks, resources
-from inductiva.molecules.simulators import GROMACS
+from inductiva.simulators import GROMACS
 from inductiva.simulators import Simulator
 from inductiva.utils.templates import (TEMPLATES_PATH,
                                        batch_replace_params_in_template)
