--- inductiva/molecules/simulators/__init__.py
@@ -1,2 +1,0 @@
-#pylint: disable=missing-module-docstring
-from .gromacs import GROMACS
